+++ seqkit/cmd/split2.go
@@ -77,11 +77,20 @@ according to the input files.
 
 		size := getFlagNonNegativeInt(cmd, "by-size")
 		parts := getFlagNonNegativeInt(cmd, "by-part")
+		lengthS := getFlagString(cmd, "by-length")
+		var length int
+		var err error
+		if lengthS != "" {
+			length, err = ParseByteSize(lengthS)
+			if err != nil {
+				checkError(fmt.Errorf("parsing sequence length: %s", lengthS))
+			}
+		}
 
 		outdir := getFlagString(cmd, "out-dir")
 		force := getFlagBool(cmd, "force")
 
-		if size == 0 && parts == 0 {
+		if size == 0 && parts == 0 && length == 0 {
 			checkError(fmt.Errorf(`one of flags should be given: -s/-p. type "seqkit split2 -h" for help`))
 		}
 
@@ -128,8 +137,10 @@ according to the input files.
 			log.Infof("split seqs from %s", source)
 			if size > 0 {
 				log.Infof("split into %d seqs per file", size)
-			} else {
+			} else if parts > 0 {
 				log.Infof("split into %d parts", parts)
+			} else {
+				log.Infof("split by sequence length: %s", lengthS)
 			}
 		}
 
@@ -186,7 +197,7 @@ according to the input files.
 					outfhs[0] = nil
 					counts = make(map[int]int, 10)
 					outfiles = make(map[int]string, 10)
-				} else {
+				} else if parts > 0 {
 					outfhs = make(map[int]*xopen.Writer, parts)
 					counts = make(map[int]int, parts)
 					outfiles = make(map[int]string, parts)
@@ -194,6 +205,11 @@ according to the input files.
 						outfhs[i] = nil
 						counts[i] = 0
 					}
+				} else {
+					outfhs = make(map[int]*xopen.Writer, 10)
+					outfhs[0] = nil
+					counts = make(map[int]int, 10)
+					outfiles = make(map[int]string, 10)
 				}
 
 				var outfh *xopen.Writer
@@ -201,6 +217,7 @@ according to the input files.
 				checkError(err)
 				i := 0 // nth part
 				j := 0 // nth record
+				n := 0 // length sum
 				for {
 					record, err = fastxReader.Read()
 					if err != nil {
@@ -223,6 +240,7 @@ according to the input files.
 						}
 						renameFileExt = false
 					}
+					n += len(record.Seq.Seq)
 
 					if size > 0 {
 						if j == size {
@@ -242,6 +260,24 @@ according to the input files.
 
 							j = 0
 						}
+					} else if length > 0 {
+						if n >= length {
+							outfhs[i].Close()
+							if !quiet {
+								log.Infof("write %d sequences to file: %s\n", counts[i], outfiles[i])
+							}
+							delete(outfhs, i)
+
+							i++
+							var outfh2 *xopen.Writer
+							outfile := filepath.Join(outdir, fmt.Sprintf("%s.part_%03d%s", filepath.Base(fileName), i+1, fileExt))
+							outfh2, err = xopen.Wopen(outfile)
+							checkError(err)
+							outfhs[i] = outfh2
+							outfiles[i] = outfile
+
+							n = 0
+						}
 					}
 
 					if outfhs[i] == nil {
@@ -260,11 +296,13 @@ according to the input files.
 
 					if size > 0 {
 						j++
-					} else {
+					} else if parts > 0 {
 						i++
 						if i == parts {
 							i = 0
 						}
+					} else {
+
 					}
 				}
 
@@ -300,6 +338,7 @@ func init() {
 	split2Cmd.Flags().StringP("read2", "2", "", "read2 file")
 	split2Cmd.Flags().IntP("by-size", "s", 0, "split sequences into multi parts with N sequences")
 	split2Cmd.Flags().IntP("by-part", "p", 0, "split sequences into N parts")
+	split2Cmd.Flags().StringP("by-length", "l", "", "split sequences into chunks of N bases, supports K/M/G suffix")
 	split2Cmd.Flags().StringP("out-dir", "O", "", "output directory (default value is $infile.split)")
 	split2Cmd.Flags().BoolP("force", "f", false, "overwrite output directory")
 }